Moved TestFlow to fixture package.
